requirements.txt
numpy
scipy
pin

# Rerun
rerun-sdk
trimesh
yourdfpy
